75505-code refactoring AB#75505
