+++ src/UKHO.ERPFacade.API/Helpers/ISapMessageBuilder.cs
@@ -1,0 +1,10 @@
+﻿using System.Xml;
+using UKHO.ERPFacade.API.Models;
+
+namespace UKHO.ERPFacade.API.Helpers
+{
+    public interface ISapMessageBuilder
+    {
+        XmlDocument BuildSapMessageXml(List<Scenario> scenarios, string traceId);
+    }
+}
