+++ src/UKHO.ERPFacade.API/Helpers/IScenarioBuilder.cs
@@ -1,0 +1,9 @@
+﻿using UKHO.ERPFacade.API.Models;
+
+namespace UKHO.ERPFacade.API.Helpers
+{
+    public interface IScenarioBuilder
+    {
+        List<Scenario> BuildScenarios(EESEvent eventData);
+    }
+}
