--- src/UKHO.ERPFacade.API/XmlHelpers/ISapMessageBuilder.cs
@@ -1,10 +1,0 @@
-﻿using System.Xml;
-using UKHO.ERPFacade.API.Models;
-
-namespace UKHO.ERPFacade.API.XmlHelpers
-{
-    public interface ISapMessageBuilder
-    {
-        XmlDocument BuildSapMessageXml(string messageTemplateName, EESEvent eesEventtData);
-    }
-}
