+++ src/UKHO.ERPFacade.Common/IO/Azure/IAzureBlobEventWriter.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json.Linq;
 
-namespace UKHO.ERPFacade.Common.IO
+namespace UKHO.ERPFacade.Common.IO.Azure
 {
     public interface IAzureBlobEventWriter
     {
