+++ src/UKHO.ERPFacade.Common/IO/Azure/IAzureTableReaderWriter.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json.Linq;
 using UKHO.ERPFacade.Common.Models.TableEntities;
 
-namespace UKHO.ERPFacade.Common.IO
+namespace UKHO.ERPFacade.Common.IO.Azure
 {
     public interface IAzureTableReaderWriter
     {
