+++ src/UKHO.ERPFacade.Common/IO/IFileSystemHelper.cs
@@ -1,0 +1,7 @@
+﻿namespace UKHO.ERPFacade.Common.IO
+{
+    public interface IFileSystemHelper
+    {
+        public bool IsFileExists(string filePath);
+    }
+}
