+++ src/UKHO.ERPFacade.Common/UKHO.ERPFacade.Common.csproj
@@ -17,6 +17,7 @@
     <PackageReference Include="Azure.Data.Tables" Version="12.8.0" />
     <PackageReference Include="Azure.Storage.Blobs" Version="12.15.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="System.IO.Abstractions" Version="19.2.18" />
   </ItemGroup>
 
 </Project>
